+++ packages/desktop-client/src/components/mobile/budget/IncomeCategoryListItem.tsx
@@ -1,0 +1,216 @@
+import { type ComponentPropsWithoutRef } from 'react';
+import { ListBoxItem } from 'react-aria-components';
+import { useTranslation } from 'react-i18next';
+
+import { Button } from '@actual-app/components/button';
+import { SvgCheveronRight } from '@actual-app/components/icons/v1';
+import { styles } from '@actual-app/components/styles';
+import { Text } from '@actual-app/components/text';
+import { theme } from '@actual-app/components/theme';
+import { View } from '@actual-app/components/view';
+import { AutoTextSize } from 'auto-text-size';
+
+import { envelopeBudget, trackingBudget } from 'loot-core/client/queries';
+import * as monthUtils from 'loot-core/shared/months';
+import { type CategoryEntity } from 'loot-core/types/models';
+
+import { useSyncedPref } from '../../../hooks/useSyncedPref';
+import { PrivacyFilter } from '../../PrivacyFilter';
+import { CellValue } from '../../spreadsheet/CellValue';
+import { useFormat } from '../../spreadsheet/useFormat';
+
+import { BudgetCell } from './BudgetCell';
+import { getColumnWidth } from './BudgetTable';
+
+type IncomeCategoryNameProps = {
+  category: CategoryEntity;
+  onEdit: (id: CategoryEntity['id']) => void;
+};
+
+function IncomeCategoryName({ category, onEdit }: IncomeCategoryNameProps) {
+  const sidebarColumnWidth = getColumnWidth({
+    isSidebar: true,
+    offset: -10,
+  });
+  return (
+    <View
+      style={{
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'flex-start',
+        width: sidebarColumnWidth,
+      }}
+    >
+      <Button
+        variant="bare"
+        style={{
+          maxWidth: sidebarColumnWidth,
+        }}
+        onPress={() => onEdit?.(category.id)}
+      >
+        <View
+          style={{
+            flexDirection: 'row',
+            alignItems: 'center',
+            justifyContent: 'flex-start',
+          }}
+        >
+          <Text
+            style={{
+              ...styles.lineClamp(2),
+              width: sidebarColumnWidth,
+              textAlign: 'left',
+              ...styles.smallText,
+            }}
+            data-testid="category-name"
+          >
+            {category.name}
+          </Text>
+          <SvgCheveronRight
+            style={{ flexShrink: 0, color: theme.tableTextSubdued }}
+            width={14}
+            height={14}
+          />
+        </View>
+      </Button>
+    </View>
+  );
+}
+
+type IncomeCategoryCellsProps = {
+  category: CategoryEntity;
+  month: string;
+  onBudgetAction: (month: string, action: string, args: unknown) => void;
+};
+
+function IncomeCategoryCells({
+  category,
+  month,
+  onBudgetAction,
+}: IncomeCategoryCellsProps) {
+  const { t } = useTranslation();
+  const format = useFormat();
+  const columnWidth = getColumnWidth();
+  const [budgetType = 'rollover'] = useSyncedPref('budgetType');
+
+  const budgeted =
+    budgetType === 'report' ? trackingBudget.catBudgeted(category.id) : null;
+
+  const balance =
+    budgetType === 'report'
+      ? trackingBudget.catSumAmount(category.id)
+      : envelopeBudget.catSumAmount(category.id);
+
+  return (
+    <View
+      style={{
+        flexDirection: 'row',
+        justifyContent: 'flex-end',
+        alignItems: 'center',
+      }}
+    >
+      {budgeted && (
+        <View
+          style={{
+            width: columnWidth,
+            justifyContent: 'center',
+            alignItems: 'flex-end',
+          }}
+        >
+          <BudgetCell
+            binding={budgeted}
+            type="financial"
+            category={category}
+            month={month}
+            onBudgetAction={onBudgetAction}
+          />
+        </View>
+      )}
+      <CellValue<'envelope-budget' | 'tracking-budget', 'sum-amount'>
+        binding={balance}
+        type="financial"
+        aria-label={t('Balance for {{categoryName}} category', {
+          categoryName: category.name,
+        })} // Translated aria-label
+      >
+        {({ type, value }) => (
+          <View>
+            <PrivacyFilter>
+              <AutoTextSize
+                key={value}
+                as={Text}
+                minFontSizePx={6}
+                maxFontSizePx={12}
+                mode="oneline"
+                style={{
+                  width: columnWidth,
+                  justifyContent: 'center',
+                  alignItems: 'flex-end',
+                  textAlign: 'right',
+                  fontSize: 12,
+                  paddingRight: 5,
+                }}
+              >
+                {format(value, type)}
+              </AutoTextSize>
+            </PrivacyFilter>
+          </View>
+        )}
+      </CellValue>
+    </View>
+  );
+}
+
+type IncomeCategoryListItemProps = ComponentPropsWithoutRef<
+  typeof ListBoxItem<CategoryEntity>
+> & {
+  month: string;
+  onEdit: (id: CategoryEntity['id']) => void;
+  onBudgetAction: (month: string, action: string, args: unknown) => void;
+};
+
+export function IncomeCategoryListItem({
+  month,
+  onEdit,
+  onBudgetAction,
+  ...props
+}: IncomeCategoryListItemProps) {
+  const { value: category } = props;
+
+  if (!category) {
+    return null;
+  }
+
+  return (
+    <ListBoxItem
+      textValue={category.name}
+      data-testid="category-row"
+      {...props}
+    >
+      <View
+        style={{
+          height: 50,
+          borderColor: theme.tableBorder,
+          flexDirection: 'row',
+          alignItems: 'center',
+          justifyContent: 'space-between',
+          paddingLeft: 5,
+          paddingRight: 5,
+          zIndex: 1,
+          borderBottomWidth: 1,
+          opacity: !!category.hidden ? 0.5 : undefined,
+          backgroundColor: monthUtils.isCurrentMonth(month)
+            ? theme.budgetCurrentMonth
+            : theme.budgetOtherMonth,
+        }}
+      >
+        <IncomeCategoryName category={category} onEdit={onEdit} />
+        <IncomeCategoryCells
+          category={category}
+          month={month}
+          onBudgetAction={onBudgetAction}
+        />
+      </View>
+    </ListBoxItem>
+  );
+}
